refactor: split workflow tasks into individual services
