+++ durable-workflow-runtime/src/main/java/com/amannmalik/workflow/runtime/task/ForkTaskService.java
@@ -1,25 +1,29 @@
 package com.amannmalik.workflow.runtime.task;
 
-import com.amannmalik.workflow.runtime.Services;
-import dev.restate.sdk.Context;
+import dev.restate.sdk.WorkflowContext;
 import dev.restate.sdk.DurableFuture;
-import dev.restate.sdk.InvocationHandle;
+import dev.restate.sdk.common.RetryPolicy;
+import dev.restate.serde.TypeTag;
 import io.serverlessworkflow.api.types.ForkTask;
 import io.serverlessworkflow.api.types.TaskItem;
-
-import java.time.Duration;
-import java.util.stream.Collectors;
 
 @dev.restate.sdk.annotation.Service
 public class ForkTaskService {
 
     @dev.restate.sdk.annotation.Handler
-    public void execute(Context ctx, ForkTask task) {
+    public void execute(WorkflowContext ctx, ForkTask task) {
         var fork = task.getFork();
-        DurableFuture.all(fork.getBranches().stream()
-                .map(TaskItem::getTask)
-                .map(subTask -> Services.invokeService(ctx, "WorkflowTaskService", "execute", subTask, Void.class, Duration.ZERO))
-                .map(InvocationHandle::attach)
-                .collect(Collectors.toUnmodifiableList())).await();
+        int i = 0;
+        java.util.List<DurableFuture<Void>> futures = new java.util.ArrayList<>();
+        for (TaskItem item : fork.getBranches()) {
+            String name = "branch-" + (i++);
+            futures.add(ctx.runAsync(name, TypeTag.of(Void.class), RetryPolicy.defaultPolicy(), () -> {
+                new WorkflowTaskService().execute(ctx, item.getTask());
+                return null;
+            }));
+        }
+        for (DurableFuture<Void> f : futures) {
+            f.await();
+        }
     }
 }
